feat: vendored decoder WBITS fix + clippy sweep + determinism audit
Three TODO items completed:

1. TODO 227 — Vendored Brotli decoder fix:
   The decoder hardcoded max_backward_distance = (1 << 24) instead of
   reading WBITS from the frame header. Vendored encoder uses WBITS=22,
   causing dictionary address extraction to fail ("invalid back-reference
   distance"). Fixed: WBITS is now threaded from the frame header through
   decode_compressed_metablock → decode_compressed_metablock_full →
   decode_compressed_metablock_full_with_trees → finish_metablock_decode.
   max_backward_distance = (1 << frame.window_bits) - WINDOW_GAP.

2. TODO 236 — Code review sweep:
   cargo clippy --fix --workspace applied. 14 files auto-fixed for:
   - Redundant field names
   - Missing backticks in docs
   - Unnecessary return values
   - Clamp-like patterns

3. TODO 237 — Determinism audit:
   - Verified: NO HashMap/HashSet in encode paths (only in tests/dict_trainer)
   - Identified risk: f32/f64 in optimal_parse cost model and FSE cost
     estimation. IEEE 754 correctly-rounded log2 should be deterministic
     across modern platforms, but not formally guaranteed.
   - All existing determinism tests pass (Brotli: compress_is_deterministic,
     ZSTD: determinism)

Remaining TODOs documented in specs 228-235 (block type switching, context
clustering, cross-block Lazy/Lazy2, FSE mode selection, shared abstractions,
streaming API). These require multi-day implementation efforts.

All 86 Brotli + 179 ZSTD tests pass (0 failures).
Version bump 0.16.10 -> 0.16.11.

+++ omnizip-flac/src/streaminfo.rs
@@ -62,7 +62,7 @@ impl StreamInfo {
         // BE value from bytes 10-12.
         let raw24 = u32::from_be_bytes([0, data[10], data[11], data[12]]);
         let sample_rate = (raw24 >> 4) & 0xFFFFF;
-        let channels = ((data[12] >> 1) & 0x07) as u8;
+        let channels = ((data[12] >> 1) & 0x07);
         let bps_high = data[12] & 0x01;
         let bps_low = data[13] >> 4;
         let bits_per_sample = (bps_high << 4) | bps_low;
